refactor : JWT 토큰 중복 메서드 제거
