+++ be/src/main/java/codesquad/issuetracker/jwt/RefreshToken.java
@@ -1,8 +1,5 @@
 package codesquad.issuetracker.jwt;
 
-import codesquad.issuetracker.exception.InvalidTokenException;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.SignatureAlgorithm;
 import java.security.Key;
@@ -35,42 +32,5 @@ public class RefreshToken extends Token {
         claims.put(ClaimKey.TOKEN_TYPE.getKeyName(), TokenType.REFRESH.name());
         claims.put(ClaimKey.MEMBER_ID.getKeyName(), memberId);
         return claims;
-    }
-
-    @Override
-    public boolean validateExpirationOfToken() {
-        try {
-            return !getClaims(token)
-                .getExpiration()
-                .before(new Date());
-        } catch (JwtException | IllegalArgumentException e) {
-            return false;
-        }
-    }
-
-    @Override
-    public String getMemberId() {
-        try {
-            return getClaims(token)
-                .get(ClaimKey.MEMBER_ID.getKeyName())
-                .toString();
-        } catch (ExpiredJwtException e) {
-            return e.getClaims().getSubject();
-        } catch (JwtException e) {
-            throw new InvalidTokenException("유효하지 않은 토큰입니다.");
-        }
-    }
-
-    @Override
-    public long getRestOfExpiration() {
-        try {
-            return getClaims(token)
-                .getExpiration()
-                .getTime() - (new Date().getTime());
-        } catch (ExpiredJwtException e) {
-            return 0;
-        } catch (JwtException e) {
-            throw new InvalidTokenException("유효하지 않은 토큰입니다.");
-        }
     }
 }
