+++ be/src/main/java/codesquad/issuetracker/jwt/Token.java
@@ -1,8 +1,12 @@
 package codesquad.issuetracker.jwt;
 
+import codesquad.issuetracker.exception.InvalidTokenException;
 import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.ExpiredJwtException;
+import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import java.security.Key;
+import java.util.Date;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -21,17 +25,45 @@ public abstract class Token {
         private final String keyName;
     }
 
-    abstract public boolean validateExpirationOfToken();
-
-    abstract public String getMemberId();
-
-    abstract public long getRestOfExpiration();
-
     protected Claims getClaims(String token) {
         return Jwts.parserBuilder()
             .setSigningKey(secretKey)
             .build()
             .parseClaimsJws(token)
             .getBody();
+    }
+
+    public boolean validateExpirationOfToken() {
+        try {
+            return !getClaims(token)
+                .getExpiration()
+                .before(new Date());
+        } catch (JwtException | IllegalArgumentException e) {
+            return false;
+        }
+    }
+
+    public String getMemberId() {
+        try {
+            return getClaims(token)
+                .get(ClaimKey.MEMBER_ID.getKeyName())
+                .toString();
+        } catch (ExpiredJwtException e) {
+            return e.getClaims().getSubject();
+        } catch (JwtException e) {
+            throw new InvalidTokenException("유효하지 않은 토큰입니다.");
+        }
+    }
+
+    public long getRestOfExpiration() {
+        try {
+            return getClaims(token)
+                .getExpiration()
+                .getTime() - (new Date().getTime());
+        } catch (ExpiredJwtException e) {
+            return 0;
+        } catch (JwtException e) {
+            throw new InvalidTokenException("유효하지 않은 토큰입니다.");
+        }
     }
 }
